+++ src/core/UniswapV3Fork.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.25;
 
-import {IERC20} from "forge-std/interfaces/IERC20.sol";
-import {ISignatureTransfer} from "permit2/src/interfaces/ISignatureTransfer.sol";
+import {IERC20} from "@forge-std/interfaces/IERC20.sol";
+import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
 import {UnsafeMath} from "../utils/UnsafeMath.sol";
 import {Panic} from "../utils/Panic.sol";
 import {SafeTransferLib} from "../vendor/SafeTransferLib.sol";
@@ -34,6 +34,7 @@ interface IUniswapV3Pool {
 
 abstract contract UniswapV3Fork is SettlerAbstract {
     using UnsafeMath for uint256;
+    using UnsafeMath for int256;
     using SafeTransferLib for IERC20;
 
     /// @dev Minimum size of an encoded swap path:
@@ -192,7 +193,7 @@ abstract contract UniswapV3Fork is SettlerAbstract {
             }
 
             {
-                int256 _buyAmount = -(zeroForOne ? amount1 : amount0);
+                int256 _buyAmount = (zeroForOne ? amount1 : amount0).unsafeNeg();
                 if (_buyAmount < 0) {
                     Panic.panic(Panic.ARITHMETIC_OVERFLOW);
                 }