+++ test/integration/UniV3CallbackPoC.t.sol
@@ -8,16 +8,16 @@ import {ISettlerActions} from "src/ISettlerActions.sol";
 import {IUniswapV3Pool} from "src/core/UniswapV3Fork.sol";
 import {AddressDerivation} from "src/utils/AddressDerivation.sol";
 
-import {ISignatureTransfer} from "permit2/src/interfaces/ISignatureTransfer.sol";
-import {IERC20} from "forge-std/interfaces/IERC20.sol";
+import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
+import {IERC20} from "@forge-std/interfaces/IERC20.sol";
 import {uniswapV3MainnetFactory} from "src/core/univ3forks/UniswapV3.sol";
 
 import {Utils} from "../unit/Utils.sol";
 import {Permit2Signature} from "../utils/Permit2Signature.sol";
 import {ActionDataBuilder} from "../utils/ActionDataBuilder.sol";
 
-import {Test, console} from "forge-std/Test.sol";
-import {MockERC20} from "solmate/src/test/utils/mocks/MockERC20.sol";
+import {Test, console} from "@forge-std/Test.sol";
+import {MockERC20} from "@solmate/test/utils/mocks/MockERC20.sol";
 
 contract UniswapV3PoolDummy {
     bytes public RETURN_DATA;