+++ test/unit/core/MakerPSMUnitTest.t.sol
@@ -3,12 +3,12 @@ pragma solidity ^0.8.25;
 
 import {MakerPSM, IPSM} from "src/core/MakerPSM.sol";
 
-import {IERC20} from "forge-std/interfaces/IERC20.sol";
+import {IERC20} from "@forge-std/interfaces/IERC20.sol";
 import {Utils} from "../Utils.sol";
 
-import {Test} from "forge-std/Test.sol";
+import {Test} from "@forge-std/Test.sol";
 
-import {ISignatureTransfer} from "permit2/src/interfaces/ISignatureTransfer.sol";
+import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
 
 contract MakerPSMDummy is MakerPSM {
     function _msgSender() internal pure override returns (address) {
@@ -27,7 +27,7 @@ contract MakerPSMDummy is MakerPSM {
         revert("unimplemented");
     }
 
-    function _dispatch(uint256, bytes4, bytes calldata) internal pure override returns (bool) {
+    function _dispatch(uint256, uint256, bytes calldata) internal pure override returns (bool) {
         revert("unimplemented");
     }
 
@@ -36,6 +36,15 @@ contract MakerPSMDummy is MakerPSM {
     }
 
     function _operator() internal pure override returns (address) {
+        revert("unimplemented");
+    }
+
+    function _permitToSellAmountCalldata(ISignatureTransfer.PermitTransferFrom calldata)
+        internal
+        pure
+        override
+        returns (uint256)
+    {
         revert("unimplemented");
     }